Fund both sides even if 24 hours have passed.
If 24 hours have passed since the last auction cleard DutchExchange
starts an auction even if only one of the sides is funded.
We want both sides funded in these cases so if getAuctionState() is
called while waiting for triggered auction to start we verify the
current side is also funded.