Change to contract buy decision logic
Fixing the query when comparing to kyber prices.
When an auction from KNC to WETH in in progress, the auction price plays
a role in this equation:

  sellTokenAmount * price.num / price.den = buyTokenAmount

DX auction price is the price for buying sellTokens using buyTokens.

In this commit we fix the market maker code so that it will query Kyber
for this rate:

  kyber.getExpectedRate(buyToken, sellToken, buyTokenAmount)

Also make changes to tests and fix many of the flaky tests.
Since dx prices change with time, some test cases still exhibit
flakiness.